requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pymupdf==1.24.14
pytesseract==0.3.10
Pillow==11.0.0
ezdxf==1.1.4
pytest==7.4.4
httpx==0.26.0
pdfplumber==0.11.0
networkx==3.2.1
shapely==2.1.0
opencv-python-headless==4.9.0.80
numpy==2.1.0
